refactor: create base entry page and add optional chaining
